Moved make_flat_button(...) and make_check_box(...) into their respective class headers
