+++ Applications/Spire/Include/Spire/Ui/Ui.hpp
@@ -58,20 +58,6 @@ namespace Spire {
   */
   QImage imageFromSvg(const QString& path, const QSize& size, const QRect& box);
 
-  //! Constructs a Spire-style CheckBox.
-  /*
-    \param label The CheckBox's text label.
-    \param parent The parent widget.
-  */
-  CheckBox* make_check_box(const QString& label, QWidget* parent);
-
-  //! Constructs a FlatButton with the default style.
-  /*!
-    \param label The button's label.
-    \param parent The parent widget.
-  */
-  FlatButton* make_flat_button(const QString& label, QWidget* parent);
-
   //! Constructs a horizontal Spire-styled table header with fixed size,
   //! immovable, column headers.
   /*
